working

+++ tests/lib/collect-artifacts.sh
@@ -43,10 +43,14 @@ coverage_after_non_nested_task() {
     if [ -d "$TESTSTMP"/coverage ] && [ $(ls "$TESTSTMP"/coverage | wc -l) -gt 0 ]; then
         local task_dir
         task_dir="$(_prepare_artifacts_path coverage-results)"
-        cp "$TESTSTMP"/coverage/* "$task_dir" || true
         pushd "$SPREAD_PATH"
-        go run "$SPREAD_PATH"/tests/utils/coverage -results-dir "$TESTSTMP"/coverage -output functions > "$task_dir"/coverage.json
+        go run "$SPREAD_PATH"/tests/utils/coverage -results-dir "$TESTSTMP"/coverage -output functions > "$task_dir"/coverage.json || true
         popd
+        if ! [ -s "$task_dir"/coverage.json ]; then
+            cp "$TESTSTMP"/coverage/* "$task_dir" || true
+            chmod 666 "$task_dir"/* || true
+            rm -f "$task_dir"/coverage.json
+        fi
     fi
 }
 
@@ -62,7 +66,12 @@ coverage_after_nested_task() {
     pushd "$SPREAD_PATH"
     go run "$SPREAD_PATH"/tests/utils/coverage -results-dir "$task_dir" -output functions > "$task_dir"/coverage.json
     popd
-    find "$task_dir" -not -name coverage.json -type f -exec rm {} \;
+    if [ -s "$task_dir"/coverage.json ]; then
+        find "$task_dir" -not -name coverage.json -type f -exec rm {} \;
+    else
+        rm "$task_dir"/coverage.json
+        chmod 666 "$task_dir"/* || true
+    fi
 }
 
 if [ "$#" == 0 ]; then
